qik-test-qa-server/ai/apps/api/requirements.txt
# FastAPI API requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
nats-py>=2.6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
